refactor(portal): separate application feature modules
